refactor: split JSON string extraction support
