+++ scripts/extract_json_strings.ts
@@ -13,228 +13,41 @@ import { dirname, normalize } from "@std/path"
 import { parse as parseJsonc } from "@std/jsonc"
 import luaparse from "npm:luaparse@0.3.1"
 
-type JsonValue = null | boolean | number | string | JsonArray | JsonObject
-type JsonArray = JsonValue[]
-type JsonObject = { [key: string]: JsonValue }
-
-type PotEntry = {
-  msgid: string
-  msgidPlural?: string
-  msgctxt?: string
-  comment?: string
-  flags: string[]
-  source: string
-}
-
-type ExtractorState = {
-  currentSourceFile: string
-  entries: PotEntry[]
-  projectName?: string
-  verbose: boolean
-  warnUnusedTypes: boolean
-  suppressWarningForFiles: Set<string>
-}
-
-type CliOptions = {
-  project?: string
-  verbose?: boolean
-  input?: string[]
-  exclude?: string[]
-  excludeDir?: string[]
-  trackedOnly?: boolean
-  output: string
-  suppress?: string[]
-  warnUnusedTypes?: boolean
-}
-
-type LuaNode = {
-  type?: string
-  name?: string
-  identifier?: { name?: string }
-  base?: LuaNode
-  expression?: LuaNode
-  arguments?: LuaNode[]
-  body?: LuaNode[]
-  comments?: LuaCommentRaw[]
-  loc?: { start?: { line?: number }; end?: { line?: number } }
-  raw?: string
-  value?: string | number | boolean | null
-  [key: string]: unknown
-}
-
-type LuaCommentRaw = {
-  value?: string
-  raw?: string
-  loc?: { start?: { line?: number }; end?: { line?: number } }
-}
-
-type LuaComment = {
-  line: number
-  text: string
-  isTranslatorComment: boolean
-  used: boolean
-}
-
-const ignorable = new Set([
-  "ascii_art",
-  "ammo_effect",
-  "behavior",
-  "charge_removal_blacklist",
-  "city_building",
-  "colordef",
-  "construction_sequence",
-  "disease_type",
-  "emit",
-  "enchantment",
-  "event_transformation",
-  "event_statistic",
-  "EXTERNAL_OPTION",
-  "hit_range",
-  "ITEM_BLACKLIST",
-  "item_group",
-  "MIGRATION",
-  "mod_tileset",
-  "monster_adjustment",
-  "MONSTER_BLACKLIST",
-  "MONSTER_FACTION",
-  "monstergroup",
-  "MONSTER_WHITELIST",
-  "mutation_type",
-  "oter_id_migration",
-  "overlay_order",
-  "overmap_connection",
-  "overmap_location",
-  "overmap_special",
-  "profession_item_substitutions",
-  "palette",
-  "region_overlay",
-  "region_settings",
-  "requirement",
-  "rotatable_symbol",
-  "SCENARIO_BLACKLIST",
-  "scent_type",
-  "skill_boost",
-  "to_cbc_migration",
-  "TRAIT_BLACKLIST",
-  "trait_group",
-  "uncraft",
-  "vehicle_group",
-  "vehicle_placement",
-  "WORLD_OPTION",
-  "sound_effect",
-])
-
-const automaticallyConvertible = new Set([
-  "achievement",
-  "activity_type",
-  "AMMO",
-  "ammunition_type",
-  "ARMOR",
-  "BATTERY",
-  "bionic",
-  "BIONIC_ITEM",
-  "BOOK",
-  "COMESTIBLE",
-  "construction_category",
-  "construction_group",
-  "CONTAINER",
-  "dream",
-  "ENGINE",
-  "event_statistic",
-  "faction",
-  "furniture",
-  "GENERIC",
-  "item_action",
-  "ITEM_CATEGORY",
-  "mutation_flag",
-  "keybinding",
-  "LOOT_ZONE",
-  "MAGAZINE",
-  "map_extra",
-  "MOD_INFO",
-  "MONSTER",
-  "morale_type",
-  "npc",
-  "npc_class",
-  "overmap_land_use_code",
-  "overmap_terrain",
-  "PET_ARMOR",
-  "score",
-  "skill",
-  "SPECIES",
-  "speech",
-  "SPELL",
-  "start_location",
-  "terrain",
-  "TOOL",
-  "TOOLMOD",
-  "TOOL_ARMOR",
-  "tool_quality",
-  "vehicle",
-  "vehicle_part",
-  "vitamin",
-  "WHEEL",
-  "help",
-  "weather_type",
-  "world_type",
-])
-
-const needsPlural = new Set([
-  "AMMO",
-  "ARMOR",
-  "BATTERY",
-  "BIONIC_ITEM",
-  "BOOK",
-  "COMESTIBLE",
-  "CONTAINER",
-  "ENGINE",
-  "GENERIC",
-  "GUN",
-  "GUNMOD",
-  "MAGAZINE",
-  "MONSTER",
-  "PET_ARMOR",
-  "SPECIES",
-  "TOOL",
-  "TOOLMOD",
-  "TOOL_ARMOR",
-  "WHEEL",
-])
-
-const useActionMessages = new Set([
-  "activate_msg",
-  "deactive_msg",
-  "out_of_power_msg",
-  "msg",
-  "menu_text",
-  "message",
-  "friendly_msg",
-  "hostile_msg",
-  "need_fire_msg",
-  "need_charges_msg",
-  "non_interactive_msg",
-  "unfold_msg",
-  "sound_msg",
-  "no_deactivate_msg",
-  "not_ready_msg",
-  "success_message",
-  "lacks_fuel_message",
-  "failure_message",
-  "descriptions",
-  "use_message",
-  "noise_message",
-  "bury_question",
-  "done_message",
-  "voluntary_extinguish_message",
-  "charges_extinguish_message",
-  "water_extinguish_message",
-  "auto_extinguish_message",
-  "activation_message",
-  "holster_msg",
-  "holster_prompt",
-  "verb",
-  "gerund",
-])
+import {
+  addContext,
+  automaticallyConvertible,
+  genderOptions,
+  gettextFunctions,
+  ignorable,
+  isArray,
+  isObject,
+  isString,
+  jsonArray,
+  jsonObjectEntries,
+  jsonStringLiteral,
+  logVerbose,
+  needsPlural,
+  objectArray,
+  parseJsonObjects,
+  poEscape,
+  poLine,
+  pythonRepr,
+  rawNameList,
+  rawTranslationString,
+  sourceName,
+  toArray,
+  useActionMessages,
+} from "./extract_json_strings_support.ts"
+import type {
+  CliOptions,
+  ExtractorState,
+  JsonObject,
+  JsonValue,
+  LuaComment,
+  LuaCommentRaw,
+  LuaNode,
+  PotEntry,
+} from "./extract_json_strings_support.ts"
 
 const foundTypes = new Set<string>()
 const warnedUnknownTypes = new Set<string>()
@@ -245,34 +58,6 @@ const memberStringPattern = new RegExp(
 )
 const arrayStringPattern = new RegExp(String.raw`^(\s*)(${jsonStringPattern})(\s*,?\s*)$`)
 const translationStringKeys = new Set(["str", "str_sp", "str_pl"])
-
-const gettextFunctions = new Map<string, { context: boolean; plural: boolean; expected: number }>([
-  ["gettext", { context: false, plural: false, expected: 1 }],
-  ["pgettext", { context: true, plural: false, expected: 2 }],
-  ["vgettext", { context: false, plural: true, expected: 3 }],
-  ["vpgettext", { context: true, plural: true, expected: 4 }],
-])
-
-const toArray = <T>(value: T | T[] | undefined): T[] => {
-  if (value === undefined) return []
-  return Array.isArray(value) ? value : [value]
-}
-
-const logVerbose = (state: ExtractorState, message: string) => {
-  if (state.verbose) console.log(message)
-}
-
-const isObject = (value: JsonValue | undefined): value is JsonObject => {
-  return typeof value === "object" && value !== null && !Array.isArray(value)
-}
-
-const isArray = (value: JsonValue | undefined): value is JsonArray => Array.isArray(value)
-
-const isString = (value: JsonValue | undefined): value is string => typeof value === "string"
-
-const sourceName = (path: string) => path.split("\\").join("/")
-
-const jsonStringLiteral = (text: string) => JSON.stringify(text)
 
 const injectJsoncTranslatorComments = (raw: string): string => {
   const out: string[] = []
@@ -328,17 +113,6 @@ const injectJsoncTranslatorComments = (raw: string): string => {
   return out.join("")
 }
 
-const poEscape = (text: string): string => {
-  return text
-    .replaceAll("\\", "\\\\")
-    .replaceAll("\t", "\\t")
-    .replaceAll("\r", "\\r")
-    .replaceAll("\n", "\\n")
-    .replaceAll('"', '\\"')
-}
-
-const poLine = (key: string, value: string): string => `${key} "${poEscape(value)}"`
-
 const writeStringBasic = (
   state: ExtractorState,
   msgid: string,
@@ -359,37 +133,6 @@ const writeStringBasic = (
     flags,
     source: state.currentSourceFile,
   })
-}
-
-const rawTranslationString = (value: JsonValue | undefined): string | undefined => {
-  if (isString(value)) return value
-  if (isObject(value)) {
-    const str = value.str ?? value.str_sp
-    return isString(str) ? str : undefined
-  }
-  return undefined
-}
-
-const pythonStringRepr = (value: string): string => {
-  const escaped = value.replaceAll("\\", "\\\\")
-  if (escaped.includes("'") && !escaped.includes('"')) return `"${escaped.replaceAll('"', '\\"')}"`
-  return `'${escaped.replaceAll("'", "\\'")}'`
-}
-
-const pythonReprQuoted = (value: JsonValue | undefined): string => {
-  if (isString(value)) return pythonStringRepr(value)
-  return pythonRepr(value)
-}
-
-const pythonRepr = (value: JsonValue | undefined): string => {
-  if (value === undefined || value === null) return "None"
-  if (isString(value)) return value
-  if (typeof value === "number") return String(value)
-  if (typeof value === "boolean") return value ? "True" : "False"
-  if (isArray(value)) return `[${value.map(pythonReprQuoted).join(", ")}]`
-  return `{${
-    Object.entries(value).map(([key, entry]) => `'${key}': ${pythonReprQuoted(entry)}`).join(", ")
-  }}`
 }
 
 const writeString = (
@@ -450,35 +193,6 @@ const extractUseActionMessages = (
       extractUseActionMessages(state, value, itemName)
     }
   }
-}
-
-const allGenders = ["f", "m", "n"]
-const genderOptions = (subject: string): string[] =>
-  allGenders.map((gender) => `${subject}:${gender}`)
-
-const addContext = (entry: JsonValue | undefined, context: string): JsonObject => {
-  if (isObject(entry)) {
-    const previous = entry.ctxt
-    return { ...entry, ctxt: isString(previous) ? `${previous}|${context}` : context }
-  }
-  return { str: entry ?? "", ctxt: context }
-}
-
-const objectArray = (value: JsonValue | undefined): JsonObject[] => {
-  if (!isArray(value)) return []
-  return value.filter(isObject)
-}
-
-const jsonArray = (value: JsonValue | undefined): JsonValue[] => isArray(value) ? value : []
-
-const jsonObjectEntries = (value: JsonValue | undefined): [string, JsonValue][] => {
-  if (!isObject(value)) return []
-  return Object.entries(value).sort(([a], [b]) => a.localeCompare(b))
-}
-
-const rawNameList = (value: JsonValue | undefined): string[] => {
-  if (!isArray(value)) return []
-  return value.map(rawTranslationString).filter((entry): entry is string => entry !== undefined)
 }
 
 const extractHarvest = (state: ExtractorState, item: JsonObject) => {
@@ -1306,22 +1020,16 @@ const extractJsonObject = (state: ExtractorState, item: JsonObject) => {
   }
 }
 
-const parseJsonFile = async (path: string): Promise<JsonValue> => {
+const parseJsonFile = async (path: string): Promise<unknown> => {
   const raw = await Deno.readTextFile(path)
-  return parseJsonc(injectJsoncTranslatorComments(raw)) as JsonValue
+  return parseJsonc(injectJsoncTranslatorComments(raw))
 }
 
 const extractJsonFile = async (state: ExtractorState, path: string) => {
   state.currentSourceFile = sourceName(path)
   logVerbose(state, `Loading ${path}`)
   const data = await parseJsonFile(path)
-  if (isArray(data)) {
-    for (const entry of data) {
-      if (isObject(entry)) extractJsonObject(state, entry)
-    }
-  } else if (isObject(data)) {
-    extractJsonObject(state, data)
-  }
+  for (const entry of parseJsonObjects(data)) extractJsonObject(state, entry)
 }
 
 const luaLiteralValue = (raw: string): string => {
